Todas as funcionalidades em um único arquivo C
Função main funciona como um "menu" em que o usuário escolhe se quer criptografar ou atacar a cifra.
Consertei um bug em que as porcentagens de letras iam no sentido contrário com o shift, se desalinhando com as porcentagens esperadas
