+++ Projeto-01_Cifra-de-Vigenere/Projeto-01_Cifra-de-Vigenere.depend
@@ -1,7 +1,9 @@
 # depslib dependency file v1.0
-1747708316 source:c:\documents\c\projects\segcomp2025.1\projeto-01_cifra-de-vigenere\main.c
+1748078077 source:c:\documents\c\projects\segcomp2025.1\projeto-01_cifra-de-vigenere\main.c
 	<stdio.h>
 	<stdlib.h>
 	<string.h>
 	<stdbool.h>
+	<ctype.h>
+	<conio.h>
