+++ Projeto-01_Cifra-de-Vigenere/Projeto-01_Cifra-de-Vigenere.layout
@@ -2,9 +2,19 @@
 <CodeBlocks_layout_file>
 	<FileVersion major="1" minor="0" />
 	<ActiveTarget name="Debug" />
-	<File name="main.c" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="6" zoom_2="0">
+	<File name="main.c" open="1" top="1" tabpos="1" split="0" active="1" splitpos="0" zoom_1="2" zoom_2="0">
 		<Cursor>
-			<Cursor1 position="3043" topLine="89" />
+			<Cursor1 position="5185" topLine="158" />
 		</Cursor>
+		<Folding>
+			<Collapse line="50" />
+			<Collapse line="58" />
+			<Collapse line="64" />
+			<Collapse line="185" />
+			<Collapse line="229" />
+			<Collapse line="248" />
+			<Collapse line="255" />
+			<Collapse line="334" />
+		</Folding>
 	</File>
 </CodeBlocks_layout_file>
