+++ Projeto-01_Cifra-de-Vigenere/ciphertext.txt
@@ -1,1 +1,1 @@
-iolpu, af tok vwwu wes xdk tlpb wdrj lsh cgdwe, qulb drmsje dolp. vv fwbv rk warq, whwrayj vae ioupq vgdvb sk qvv eowuwh. af tok owiv dcivdqldf dkr fkb jasbgk', hrogisje' ofo qofyhoh' vqodwv lvdh oddhwrq fidkpg, lsh hldhbi ppmiq. fi idp yaw pmf qvv pwc sda r eiotv ydrf qvjaak ghho ytv hyaxirvdv. yp agvi wilpqqwgy qd sljhvwi.
+hokpu, ze soj vwwu wds xdj tlpa wcri ksh bfcwd, quka crlsid cokp. vv fwav rj wzrp, whwrzyj vzd houpq vfcva sj pvv enwuwh. ze soj owhv dbhvcqkce djr eka izsafj', groghsid' oeo poeyhng' vpocwv kvch occhwrp fhcjpg, ksh gkchah pplhp. eh hdo yzw ple pvv pwb sdz r ehotv ycre pvizzj fghn ytv gyzxhrvcv. yp zfvh whkpqpwfy pc sligvwi.
